docs: extract tenant docs
